+++ src/lib/index.ts
@@ -1,0 +1,12 @@
+/**
+ * The MIT License (MIT)
+ * Copyright (c) 2017 Forfuture, LLC <we@forfuture.co.ke>
+ */
+
+import * as project from "./project";
+import * as versions from "./versions";
+
+export {
+    project,
+    versions,
+};